app/server/requirements.txt
bcrypt
cryptography
python-dotenv
